Nickname management refactoring

+++ server/src/logic/nickname.ts
@@ -3,7 +3,8 @@ import LogicError                                 from "./LogicError"
 
 import { ResultSetHeader                        } from "mysql2"
 import { Connection, FieldPacket, RowDataPacket } from "mysql2/promise"
-import { Logger                                 } from "winston"
+import { Logger, add                                 } from "winston"
+import { hasWs                                  } from "util/string"
 import { UserManager, User, UserInfo            } from "./user"
 
 export interface CreationOptions {
@@ -12,26 +13,35 @@ export interface CreationOptions {
     readonly logger?:     Logger
 }
 
-export interface SetUserNicknamesOptions {
-    throwOnInvalidUser?: boolean
-    throwOnLimit?:       boolean
+export interface GetNicknameOwnerIdOptions {
+    throwOnInvalidNickname?: boolean
+    throwOnOwnerNotFound?:   boolean
 }
 
-export interface DeleteUserNicknameOptions {
-    throwOnInvalidUser?:     boolean
-    throwOnInvalidNickname?: boolean
-}
+export interface GetNicknameOwnerInfoOptions extends GetNicknameOwnerIdOptions {}
 
 export interface AddUserNicknameOptions {
+    throwOnInvalidNickname?: boolean
+    throwOnInvalidUser?:     boolean
+    throwOnLimit?:           boolean
+    throwOnDuplicate?:       boolean
+}
+
+export interface SetUserNicknamesOptions extends AddUserNicknameOptions {}
+
+export interface DeleteNicknameOptions {
+    throwOnInvalidNickname?:  boolean
+    throwOnNicknameNotFound?: boolean
+}
+
+export interface DeleteUserNicknameOptions extends DeleteNicknameOptions {
     throwOnInvalidUser?: boolean
-    throwOnLimit?:       boolean
-    throwOnDuplicate?:   boolean
 }
 
 export interface NicknameManager {
-    forceSetUserNicknames(connection: Connection, user: User, nicknames: string[] | null): Promise<void>
+    forceSetUserNicknames(connection: Connection, user: User, nicknames: string[] | null): Promise<number>
 
-    setUserNicknames(connection: Connection, user: User, nicknames: string[] | null, options?: SetUserNicknamesOptions): Promise<boolean>
+    setUserNicknames(connection: Connection, user: User, nicknames: string[] | null, options?: SetUserNicknamesOptions): Promise<number>
 
 
     forceDeleteAllUserNicknames(connection: Connection, user: User): Promise<void>
@@ -49,20 +59,17 @@ export interface NicknameManager {
 
     forceDeleteNickname(connection: Connection, nickname: string): Promise<void>
 
-    deleteNickname(connection: Connection, nickname: string, throwOnFailure:  true):    Promise<true>
-    deleteNickname(connection: Connection, nickname: string, throwOnFailure?: boolean): Promise<boolean>
+    deleteNickname(connection: Connection, nickname: string, options?: DeleteNicknameOptions): Promise<boolean>
 
 
     forceGetNicknameOwnerInfo(connection: Connection, nickname: string): Promise<UserInfo>
 
-    getNicknameOwnerInfo(connection: Connection, nickname: string, throwOnFailure:  true):    Promise<UserInfo>
-    getNicknameOwnerInfo(connection: Connection, nickname: string, throwOnFailure?: boolean): Promise<UserInfo | undefined>
+    getNicknameOwnerInfo(connection: Connection, nickname: string, options?: GetNicknameOwnerInfoOptions): Promise<UserInfo | undefined>
 
 
     forceGetNicknameOwnerId(connection: Connection, nickname: string): Promise<number>
 
-    getNicknameOwnerId(connection: Connection, nickname: string, throwOnFailure:  true):    Promise<number>
-    getNicknameOwnerId(connection: Connection, nickname: string, throwOnFailure?: boolean): Promise<number | undefined>
+    getNicknameOwnerId(connection: Connection, nickname: string, options?: GetNicknameOwnerIdOptions): Promise<number | undefined>
 
 
     forceGetUserNicknameCount(connection: Connection, user: User): Promise<number>
@@ -91,14 +98,16 @@ export class DefaultNicknameManager implements NicknameManager {
         this.logger      = options.logger
     }
 
-    async forceSetUserNicknames(connection: Connection, user: User, nicknames: string[] | null) {
-        await this.setUserNicknames(connection, user, nicknames, {
-            throwOnLimit:       true,
-            throwOnInvalidUser: true
+    async forceSetUserNicknames(connection: Connection, user: User, nicknames: string[] | null): Promise<number> {
+        return await this.setUserNicknames(connection, user, nicknames, {
+            throwOnInvalidNickname: true,
+            throwOnDuplicate:       true,
+            throwOnLimit:           true,
+            throwOnInvalidUser:     true
         })
     }
 
-    async setUserNicknames(connection: Connection, user: User, nicknames: string[] | null, options?: SetUserNicknamesOptions): Promise<boolean> {
+    async setUserNicknames(connection: Connection, user: User, nicknames: string[] | null, options: SetUserNicknamesOptions = {}): Promise<number> {
         this.logger?.debug(typeof user === "string" ? `Setting nicknames of user "${user}" to [${nicknames?.join(", ")}]...`
                                                     : `Setting nicknames of user with id ${user} to [${nicknames?.join(", ")}]...`)
 
@@ -112,22 +121,33 @@ export class DefaultNicknameManager implements NicknameManager {
 
             this.logger?.debug(message)
 
-            return false
+            return 0
         }
 
         const userId = await this.userManager.getUserId(connection, user, options?.throwOnInvalidUser)
 
         if (userId == null)
-            return false
+            return 0
 
         await this.deleteAllUserNicknames(connection, userId)
 
+        const { throwOnInvalidNickname, throwOnDuplicate } = options
+
+        const addOptions = { throwOnDuplicate, throwOnInvalidNickname }
+
+        let setCount = 0
+
         if (count > 0)
-            await Promise.all(nicknames!.map(nickname => this.addUserNickname(connection, userId, nickname, { throwOnDuplicate: true })))
+            await Promise.all(nicknames!.map(async nickname => {
+                const added = await this.addUserNickname(connection, userId, nickname, addOptions)
+
+                if (added)
+                    ++setCount
+            }))
 
         this.logger?.debug("Set")
 
-        return true
+        return setCount
     }
 
     async forceDeleteAllUserNicknames(connection: Connection, user: User) {
@@ -164,14 +184,20 @@ export class DefaultNicknameManager implements NicknameManager {
 
     async forceDeleteUserNickname(connection: Connection, user: User, nickname: string) {
         await this.deleteUserNickname(connection, user, nickname, {
-            throwOnInvalidNickname: true,
-            throwOnInvalidUser:     true
+            throwOnInvalidNickname:  true,
+            throwOnInvalidUser:      true,
+            throwOnNicknameNotFound: true
         })
     }
 
     async deleteUserNickname(connection: Connection, user: User, nickname: string, options: DeleteUserNicknameOptions): Promise<boolean> {
+        nickname = normNickname(nickname)
+
         this.logger?.debug(typeof user === "string" ? `Deleting nickname "${nickname}" of user "${user}"...`
                                                     : `Deleting all nickname "${nickname}" of user with id ${user}...`)
+
+        if (checkNickname(nickname, options.throwOnInvalidNickname, this.logger) != null)
+            return false
 
         const id = await this.userManager.getUserId(connection, user, options.throwOnInvalidUser)
 
@@ -198,13 +224,19 @@ export class DefaultNicknameManager implements NicknameManager {
     }
 
     async forceDeleteNickname(connection: Connection, nickname: string) {
-        await this.deleteNickname(connection, nickname, true)
+        await this.deleteNickname(connection, nickname, {
+            throwOnInvalidNickname:  true,
+            throwOnNicknameNotFound: true
+        })
     }
 
-    async deleteNickname(connection: Connection, nickname: string, throwOnFailure:  true):            Promise<true>
-    async deleteNickname(connection: Connection, nickname: string, throwOnFailure?: boolean):         Promise<boolean>
-    async deleteNickname(connection: Connection, nickname: string, throwOnFailure:  boolean = false): Promise<boolean> {
+    async deleteNickname(connection: Connection, nickname: string, options?: DeleteNicknameOptions)  : Promise<boolean> {
+        nickname = normNickname(nickname)
+
         this.logger?.debug(`Deleting nickname "${nickname}"...`)
+
+        if (checkNickname(nickname, options?.throwOnInvalidNickname, this.logger) != null)
+            return false
 
         const [result] = await connection.execute("DELETE FROM Nicknames WHERE nickname = ?", [nickname]) as [ResultSetHeader, FieldPacket[]]
         const deleted  = result.affectedRows !== 0
@@ -212,7 +244,7 @@ export class DefaultNicknameManager implements NicknameManager {
         if (!deleted) {
             const message = `Nickname "${nickname}" not found`
 
-            if (throwOnFailure)
+            if (options?.throwOnNicknameNotFound)
                 throw new LogicError(message)
 
             this.logger?.debug(message)
@@ -226,41 +258,44 @@ export class DefaultNicknameManager implements NicknameManager {
     }
 
     async forceGetNicknameOwnerInfo(connection: Connection, nickname: string): Promise<UserInfo> {
-        return await this.getNicknameOwnerInfo(connection, nickname, true)
+        return (await this.getNicknameOwnerInfo(connection, nickname, {
+            throwOnInvalidNickname: true,
+            throwOnOwnerNotFound:   true
+        }))!
     }
 
-    async getNicknameOwnerInfo(connection: Connection, nickname: string, throwOnFailure:  true):            Promise<UserInfo>
-    async getNicknameOwnerInfo(connection: Connection, nickname: string, throwOnFailure?: boolean):         Promise<UserInfo | undefined>
-    async getNicknameOwnerInfo(connection: Connection, nickname: string, throwOnFailure:  boolean = false): Promise<UserInfo | undefined> {
+    async getNicknameOwnerInfo(connection: Connection, nickname: string, options?: GetNicknameOwnerInfoOptions): Promise<UserInfo | undefined> {
         this.logger?.debug(`Getting nickname "${nickname}" owner info...`)
 
-        const id = await this.getNicknameOwnerId(connection, nickname, throwOnFailure)
+        const id = await this.getNicknameOwnerId(connection, nickname, options)
 
         if (id == null)
             return undefined
 
-        const info = await this.userManager.getUserInfo(connection, id)
-
-        this.logger?.debug("Got")
-
-        return info
+        return await this.userManager.getUserInfo(connection, id)
     }
 
     async forceGetNicknameOwnerId(connection: Connection, nickname: string): Promise<number> {
-        return await this.getNicknameOwnerId(connection, nickname, true)
+        return (await this.getNicknameOwnerId(connection, nickname, {
+            throwOnInvalidNickname: true,
+            throwOnOwnerNotFound:   true
+        }))!
     }
 
-    async getNicknameOwnerId(connection: Connection, nickname: string, throwOnFailure:  true):            Promise<number>
-    async getNicknameOwnerId(connection: Connection, nickname: string, throwOnFailure?: boolean):         Promise<number | undefined>
-    async getNicknameOwnerId(connection: Connection, nickname: string, throwOnFailure:  boolean = false): Promise<number | undefined> {
+    async getNicknameOwnerId(connection: Connection, nickname: string, options?: GetNicknameOwnerIdOptions): Promise<number | undefined> {
+        nickname = normNickname(nickname)
+
         this.logger?.debug(`Getting nickname "${nickname}" owner id...`)
+
+        if (checkNickname(nickname, options?.throwOnInvalidNickname, this.logger) != null)
+            return
 
         const [rows] = await connection.execute("SELECT user_id FROM Nicknames WHERE nickname = ?", [nickname]) as [RowDataPacket[], FieldPacket[]]
 
         if (rows.length === 0) {
             const message = `There is no owner of nickname "${nickname}"`
 
-            if (throwOnFailure)
+            if (options?.throwOnOwnerNotFound)
                 throw new LogicError(message)
         
             this.logger?.debug(message)
@@ -285,10 +320,8 @@ export class DefaultNicknameManager implements NicknameManager {
 
         const id = await this.userManager.getUserId(connection, user, throwOnFailure)
 
-        if (id == null) {
-            this.logger?.debug("User not found")
+        if (id == null)
             return 0
-        }
 
         const [rows] = await connection.execute("SELECT COUNT(*) AS count FROM Nicknames WHERE user_id = ?", [id]) as [RowDataPacket[], FieldPacket[]]
         const count  = rows[0].count
@@ -300,15 +333,21 @@ export class DefaultNicknameManager implements NicknameManager {
 
     async forceAddUserNickname(connection: Connection, user: User, nickname: string) {
         await this.addUserNickname(connection, user, nickname, {
-            throwOnInvalidUser: true,
-            throwOnLimit:       true,
-            throwOnDuplicate:   true
+            throwOnInvalidUser:     true,
+            throwOnInvalidNickname: true,
+            throwOnLimit:           true,
+            throwOnDuplicate:       true
         })
     }
 
     async addUserNickname(connection: Connection, user: User, nickname: string, options?: AddUserNicknameOptions): Promise<boolean> {
-        this.logger?.debug(typeof user === "string" ? `Adding nickname "${nickname}" of user "${user}"...`
-                                                    : `Adding nickname "${nickname}" of user with id ${user}...`)
+        nickname = normNickname(nickname)
+
+        this.logger?.debug(typeof user === "string" ? `Adding nickname "${nickname}" to user "${user}"...`
+                                                    : `Adding nickname "${nickname}" to user with id ${user}...`)
+
+        if (checkNickname(nickname, options?.throwOnInvalidNickname, this.logger) != null)
+            return false
 
         const id = await this.userManager.getUserId(connection, user, options?.throwOnInvalidUser)
 
@@ -369,4 +408,38 @@ export class DefaultNicknameManager implements NicknameManager {
 
         return nicknames
     }
+}
+
+export function checkNickname(nickname: string, throwOnFailure: boolean = false, logger?: Logger): string | undefined {
+    const invalidReason = validateNickname(nickname)
+
+    if (invalidReason != null) {
+        if (throwOnFailure)
+            throw new LogicError(invalidReason)
+
+        logger?.debug(invalidReason)
+    }
+
+    return invalidReason
+}
+
+export function validateNickname(nickname: string): string | undefined {
+    const MIN_LENGTH = 4
+
+    if (nickname.length < MIN_LENGTH)
+        return `Nickname "${nickname}" is too short. Minimum ${MIN_LENGTH} characters required`
+
+    const MAX_LENGTH = 255
+
+    if (nickname.length > MAX_LENGTH)
+        return `Nickname is too long. Maximum ${MAX_LENGTH} characters allowed`
+
+    if (hasWs(nickname))
+        return `Nickname "${nickname}" contains whitespace`
+
+    return undefined
+}
+
+export function normNickname(nickname: string): string {
+    return nickname.trim()
 }
